+++ lib/Popups/home_create_new_popup.dart
@@ -1,12 +1,9 @@
 import 'dart:io';
-
 import 'package:flutter/material.dart';
 import 'package:memories_photos/Editor/collage_maker_page.dart';
-import 'package:memories_photos/Editor/editor_page.dart';
 import 'package:memories_photos/Pages/More/camera_timelapse_page.dart';
-import 'package:memories_photos/Popups/new_album_popup.dart';
 import 'package:memories_photos/Popups/toast.dart';
-import 'package:memories_photos/image_helper.dart';
+import 'package:memories_photos/Scripts/image_helper.dart';
 import 'package:memories_photos/settings.dart';
 
 Future<bool> showHomeCreateNewPopup(BuildContext context) async {
@@ -29,16 +26,18 @@ Future<bool> showHomeCreateNewPopup(BuildContext context) async {
               Expanded(
                 child: OutlinedButton.icon(
                   label: Text(
-                    "Record Video",
+                    "Timelapse",
                     style: TextStyle(
                       fontFamily: Settings.ElmsSans
                     ),
                   ),
-                  icon: Icon(Icons.video_camera_front_outlined),
-                  onPressed: () {
-                    // TODO: recording video
+                  icon: Icon(Icons.timelapse_outlined),
+                  onPressed: () async {
                     returnTrue = true;
                     Navigator.of(context).pop();
+                    if (Platform.isWindows) // TODO: Implement Camera & Timelapse for windows
+                      showStyledToast("Coming Soon!", context);
+                    else Navigator.of(context).push(MaterialPageRoute(builder: (context) => CameraTakeTimelapsePage()));
                   },
                 ),
               ),
@@ -63,23 +62,6 @@ Future<bool> showHomeCreateNewPopup(BuildContext context) async {
             ],
           ),
           SizedBox(height: 5),
-          FilledButton.icon(
-            label: Text(
-              "Timelapse",
-              style: TextStyle(
-                fontFamily: Settings.ElmsSans
-              ),
-            ),
-            icon: Icon(Icons.timelapse_outlined),
-            onPressed: () async {
-              returnTrue = true;
-              Navigator.of(context).pop();
-              if (Platform.isWindows) // TODO: Implement for windows
-                showStyledToast("Coming Soon!", context);
-              else Navigator.of(context).push(MaterialPageRoute(builder: (context) => CameraTakeTimelapsePage()));
-            },
-          ),
-          SizedBox(height: 5),
           Text(
             "🌲 Create New ✨",
             textAlign: .center,
@@ -89,16 +71,6 @@ Future<bool> showHomeCreateNewPopup(BuildContext context) async {
             ),
           ),
           ListTile(
-            title: Text("Album"),
-            shape: RoundedRectangleBorder(borderRadius: .circular(25 * Settings.rm)),
-            leading: Icon(Icons.photo_album_outlined),
-            onTap: () async {
-              await showNewAlbumDialog(context);
-              returnTrue = true;
-              Navigator.of(context).pop();
-            },
-          ),
-          ListTile(
             title: Text("Collage"),
             shape: RoundedRectangleBorder(borderRadius: .circular(25 * Settings.rm)),
             leading: Icon(Icons.collections_outlined),
@@ -106,16 +78,6 @@ Future<bool> showHomeCreateNewPopup(BuildContext context) async {
               returnTrue = true;
               Navigator.of(context).pop();
               Navigator.of(context).push(MaterialPageRoute(builder: (context) => CollageMakerPage(),));
-            },
-          ),
-          ListTile(
-            title: Text("Edit then save as"),
-            shape: RoundedRectangleBorder(borderRadius: .circular(25 * Settings.rm)),
-            leading: Icon(Icons.image_outlined),
-            onTap: () {
-              returnTrue = true;
-              Navigator.of(context).pop();
-              Navigator.of(context).push(MaterialPageRoute(builder: (context) => EditorPage(),));
             },
           ),
           ListTile(